final update before going live (famous last words)
- set unique surrogate key for vercel-hosted content being served
- pay no attention to the insanity that now will become fastly.toml.
- set surrogate key for vercel content
- disable font: false
- strip non-EN for now

+++ build-all.sh
@@ -21,7 +21,10 @@ MINIFY=$(command -v minify || echo "./compute-js/bin/minify")
 # minify everything
 test -d build/minified || $MINIFY -r --sync --preserve=all -o build/minified build/site
 
-# remove locales from minified (Fastly) site
-for locale in $(yq -r '.plugins[] | select(type == "object" and has("i18n")) | .i18n.languages[].locale' < mkdocs.yml | grep -v 'en'); do
-  rm -fr "build/minified/site/$locale"
-done
+STRIP_NONENGLISH_LOCALES=${STRIP_LOCALES:-true}
+if ${STRIP_NONENGLISH_LOCALES}; then
+  # remove locales from minified (Fastly) site
+  for locale in $(yq -r '.plugins[] | select(type == "object" and has("i18n")) | .i18n.languages[].locale' < mkdocs.yml | grep -v 'en'); do
+    rm -fr "build/minified/site/$locale"
+  done
+fi
